+++ src/activities/util/FullScreenMessageActivity.h
@@ -15,7 +15,7 @@ class FullScreenMessageActivity final : public Activity {
  public:
   explicit FullScreenMessageActivity(GfxRenderer& renderer, MappedInputManager& mappedInput, std::string text,
                                      const EpdFontFamily::Style style = EpdFontFamily::REGULAR,
-                                     const HalDisplay::RefreshMode refreshMode = HalDisplay::FAST_REFRESH)
+                                     const HalDisplay::RefreshMode refreshMode = HalDisplay::FULL_REFRESH)
       : Activity("FullScreenMessage", renderer, mappedInput),
         text(std::move(text)),
         style(style),